Move customtexture initialization to the mod system. Split off texture dumping since it's a rather separate feature that doesn't share any code with the loader.
