+++ libswirl/rend/gles/texture_dump.cpp
@@ -16,10 +16,7 @@
 	 You should have received a copy of the GNU General Public License
 	 along with reicast.  If not, see <https://www.gnu.org/licenses/>.
  */
-#include "CustomTexture.h"
-
-#include <algorithm>
-#include <sstream>
+#include "texture_dump.h"
 #include <sys/types.h>
 #include <sys/stat.h>
 #ifdef _MSC_VER
@@ -31,141 +28,16 @@
 #include "deps/libpng/png.h"
 #include "reios/reios.h"
 
-#include "scripting/mods.h"
+#include "mods/mods.h"
 
-void CustomTexture::LoaderThread()
-{
-	while (initialized)
-	{
-		TextureCacheData *texture;
-		
-		do {
-			texture = NULL;
+#include <sstream>
 
-			work_queue_mutex.Lock();
-			if (!work_queue.empty())
-			{
-				texture = work_queue.back();
-				work_queue.pop_back();
-			}
-			work_queue_mutex.Unlock();
-			
-			if (texture != NULL)
-			{
-				texture->ComputeHash();
-				if (texture->custom_image_data != NULL)
-				{
-					delete [] texture->custom_image_data;
-					texture->custom_image_data = NULL;
-				}
-				if (!texture->dirty)
-				{
-					int width, height;
-					u8 *image_data = LoadCustomTexture(texture->texture_hash, width, height);
-					if (image_data == NULL)
-					{
-						image_data = LoadCustomTexture(texture->old_texture_hash, width, height);
-					}
-					if (image_data != NULL)
-					{
-						texture->custom_width = width;
-						texture->custom_height = height;
-						texture->custom_image_data = image_data;
-					}
-				}
-				texture->custom_load_in_progress--;
-			}
-
-		} while (texture != NULL);
-		
-		wakeup_thread.Wait();
-	}
-}
-
-std::string CustomTexture::GetGameId()
-{
-   std::string game_id = reios_product_number;
-   const size_t str_end = game_id.find_last_not_of(" ");
-   if (str_end == std::string::npos)
-	  return "";
-   game_id = game_id.substr(0, str_end + 1);
-   std::replace(game_id.begin(), game_id.end(), ' ', '_');
-
-   return game_id;
-}
-
-bool CustomTexture::Init()
-{
-	if (!initialized)
-	{
-		initialized = true;
-#ifndef TARGET_NO_THREADS
-		std::string game_id = GetGameId();
-		if (game_id.length() > 0)
-		{
-			textures_path = get_readonly_data_path(DATA_PATH) + "textures/" + game_id + "/";
-
-			DIR *dir = opendir(textures_path.c_str());
-			if (dir != NULL)
-			{
-				printf("Found custom textures directory: %s\n", textures_path.c_str());
-				custom_textures_available = true;
-				closedir(dir);
-				loader_thread.Start();
-			}
-		}
-#endif
-	}
-	return custom_textures_available;
-}
-
-void CustomTexture::Terminate()
-{
-	if (initialized)
-	{
-		initialized = false;
-		work_queue_mutex.Lock();
-		work_queue.clear();
-		work_queue_mutex.Unlock();
-		wakeup_thread.Set();
-#ifndef TARGET_NO_THREADS
-		loader_thread.WaitToEnd();
-#endif
-	}
-}
-
-u8* CustomTexture::LoadCustomTexture(u32 hash, int& width, int& height)
-{
-	if (unknown_hashes.find(hash) != unknown_hashes.end())
-		return NULL;
-	std::stringstream path;
-	path << textures_path << std::hex << hash << ".png";
-
-	u8 *image_data = loadPNGData(path.str(), width, height);
-	if (image_data == NULL)
-		unknown_hashes.insert(hash);
-
-	return image_data;
-}
-
-void CustomTexture::LoadCustomTextureAsync(TextureCacheData *texture_data)
-{
-	if (!Init())
-		return;
-
-	texture_data->custom_load_in_progress++;
-	work_queue_mutex.Lock();
-	work_queue.insert(work_queue.begin(), texture_data);
-	work_queue_mutex.Unlock();
-	wakeup_thread.Set();
-}
-
-void CustomTexture::DumpTexture(u32 hash, int w, int h, GLuint textype, void *temp_tex_buffer)
+void gles_dump_texture(u32 hash, int w, int h, GLuint textype, void *temp_tex_buffer)
 {
 	std::string base_dump_dir = get_writable_data_path(DATA_PATH "texdump/");
 	if (!file_exists(base_dump_dir))
 		make_directory(base_dump_dir);
-	std::string game_id = GetGameId();
+	std::string game_id = get_game_id();
 	if (game_id.length() == 0)
 	   return;
